refactor insurance personas into separate module
