+++ src/modules/insurance/personas.ts
@@ -1,0 +1,10 @@
+export const insurancePersonaItems = [
+  { key: "consumer", label: "개인고객", icon: "🙋" },
+  { key: "agent", label: "설계사", icon: "💼" },
+  { key: "expert", label: "전문가", icon: "🧠" },
+  { key: "corporate", label: "기업·단체", icon: "🏢" },
+  { key: "licensePrep", label: "취업·수험", icon: "📚" },
+  { key: "overseas", label: "해외/유학생", icon: "🌏" },
+];
+
+export default insurancePersonaItems;
